# docs/swagger.yaml
definitions:
  entity.Cluster:
    properties:
      Status:
        type: string
      api_server:
        type: string
      createdAt:
        type: string
      id:
        type: string
      kube_config:
        type: string
      name:
        type: string
      token:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
      version:
        type: string
    type: object
...
  entity.Instance:
    properties:
      cluster:
        $ref: '#/definitions/model.Cluster'
      cluster_id:
        type: string
      count:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      limit_cpu:
        type: string
      limit_memory:
        type: string
      name:
        type: string
      namespace:
        type: string
      request_cpu:
        type: string
      request_memory:
        type: string
      secret:
        items:
          $ref: '#/definitions/entity.SecretInfo'
        type: array
      service_info:
        $ref: '#/definitions/entity.ServiceInfo'
      service_type:
        type: string
      status:
        type: string
      template:
        $ref: '#/definitions/model.Templates'
      template_id:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
      volume:
        type: string
      yaml:
        type: string
    type: object
  entity.InstancePage:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.Instance'
        type: array
      total:
        type: integer
    type: object
...
  entity.Postgres:
    properties:
      cluster:
        $ref: '#/definitions/entity.Cluster'
      cluster_id:
        type: string
      name:
        type: string
      namespace:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  entity.SecretInfo:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
  entity.ServiceAddr:
    properties:
      host:
        type: string
      port:
        type: integer
    type: object
  entity.ServiceInfo:
    properties:
      addresses:
        items:
          $ref: '#/definitions/entity.ServiceAddr'
        type: array
      service_type:
        type: string
    type: object
...
  model.Cluster:
    properties:
      Status:
        type: string
      api_server:
        type: string
      createdAt:
        type: string
      id:
        type: string
      kube_config:
        type: string
      name:
        type: string
      token:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      version:
        type: string
    type: object
  model.Templates:
    properties:
      advance_template:
        type: string
      base_template:
        type: string
      createdAt:
        type: string
      icon:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/postgres:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 实例名
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
      tags:
      - 删除实例
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: postgresql实例名称
        in: query
        name: instance_name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Instance'
      tags:
      - 根据postgresql名称获取postgres对象
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Instance'
      responses:
        "200":
          description: OK
      tags:
      - 更新实例
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Instance'
      responses:
        "200":
          description: OK
      tags:
      - 创建postgres实例
  /api/v1/postgres/:postgres_id:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: postgresql实例ID
        in: query
        name: postgres_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Instance'
      tags:
      - 根据postgresql ID获取postgres对象
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Instance'
      responses:
        "200":
          description: OK
      tags:
      - 编辑实例
  /api/v1/postgres/:postgres_id/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: postgresql实例ID
        in: query
        name: postgres_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      tags:
      - 根据postgresql ID获取postgres状态
  /api/v1/postgres/template:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Postgres'
      responses:
        "200":
          description: OK
      tags:
      - 从模板创建postgres实例
  /api/v1/postgressqls:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 长度
        in: query
        name: size
        required: true
        type: string
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.InstancePage'
      tags:
      - 根据用户ID获取postgres实例列表
...
  /api/ws/v1/pod/log:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 集群ID
        in: query
        name: cluster_id
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: pod名称
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 获取Pod的日志
swagger: "2.0"